Proper loading/updating/switching of MV's USD & Override files.
This commit adds proper loading of overrides, in particular, it creates a
loader which is aware that the usd overrides has to go into a
mvUsdCompoundShape's path stack. Upon update (of either of the original
mvUsd or override), it will preserve the order and update just the
updated stack.

For this, we need to get the previous version's path, and search for it
in the current compound shape, and the replace it.

This commit is full of debug info which is added to aid development. It
is left in this commit for future reference, but will be removed in the
next commit.